+++ emerald-core/src/keystore/serialize/mod.rs
@@ -6,7 +6,7 @@ mod byte_array;
 mod crypto;
 mod error;
 
-use self::address::try_extract_address;
+pub use self::address::try_extract_address;
 pub use self::crypto::{CoreCrypto, Iv, Mac, Salt, decode_str};
 pub use self::error::Error;
 use super::{CIPHER_IV_BYTES, Cipher, CryptoType, KDF_SALT_BYTES, Kdf, KeyFile};
@@ -135,66 +135,6 @@ impl KeyFile {
 
         Ok(kf)
     }
-
-    /// Serializes into JSON file with the name format `UTC--<timestamp>Z--<uuid>`
-    ///
-    /// # Arguments
-    ///
-    /// * `dir` - path to destination directory
-    /// * `addr` - a public address (optional)
-    ///
-    pub fn flush<P: AsRef<Path>>(&self, dir: P, filename: Option<&str>) -> Result<(), Error> {
-        let tmp;
-        let name = match filename {
-            Some(n) => n,
-            None => {
-                tmp = generate_filename(&self.uuid.to_string());
-                &tmp
-            }
-        };
-        let path = dir.as_ref().join(name);
-
-        Ok(write(self, path)?)
-    }
-
-    /// Search of `KeyFile` by specified `Address`
-    /// Returns set of filepath and `Keyfile`
-    ///
-    /// # Arguments
-    ///
-    /// * `path` - path with keystore files
-    ///
-    pub fn search_by_address<P: AsRef<Path>>(
-        addr: &Address,
-        path: P,
-    ) -> Result<(PathBuf, KeyFile), Error> {
-        let entries = fs::read_dir(path)?;
-
-        for entry in entries {
-            let path = entry?.path();
-
-            if path.is_dir() {
-                continue;
-            }
-
-            let mut file = fs::File::open(&path)?;
-            let mut content = String::new();
-
-            if file.read_to_string(&mut content).is_err() {
-                continue;
-            }
-
-            match try_extract_address(&content) {
-                Some(a) if a == *addr => {
-                    let kf = KeyFile::decode(content)?;
-                    return Ok((path.to_owned(), kf));
-                }
-                _ => continue,
-            }
-        }
-
-        Err(Error::NotFound)
-    }
 }
 
 impl Encodable for KeyFile {
@@ -210,127 +150,6 @@ impl Encodable for KeyFile {
 
         }
     }
-}
-
-
-/// Writes out `Keyfile` into disk accordingly to `p` path
-/// Try to match one wvariant of KeyFile:
-///     `..Core` - for normal `Keyfile` created as JSON safe storage
-///     `..HD` - for usage with HD wallets
-///
-/// #Arguments:
-/// kf - `Keyfile` to be written
-/// p - destination route (path + filename)
-///
-pub fn write<P: AsRef<Path>>(kf: &KeyFile, p: P) -> Result<(), Error> {
-    let json = json::encode(&kf)?;
-
-    let mut file = File::create(&p)?;
-    file.write_all(json.as_ref()).ok();
-
-    Ok(())
-}
-
-/// Lists addresses for `Keystore` files in specified folder.
-/// Can include hidden files if flag set.
-///
-/// # Arguments
-///
-/// * `path` - target directory
-/// * `showHidden` - flag to show hidden `Keystore` files
-///
-/// # Return:
-/// Array of tuples (name, address, description, is_hidden)
-///
-pub fn list_accounts<P: AsRef<Path>>(
-    path: P,
-    show_hidden: bool,
-) -> Result<Vec<(String, String, String, bool)>, Error> {
-    let mut accounts: Vec<(String, String, String, bool)> = vec![];
-    for e in read_dir(&path)? {
-        if e.is_err() {
-            continue;
-        }
-        let entry = e.unwrap();
-        let mut content = String::new();
-        if let Ok(mut keyfile) = File::open(entry.path()) {
-            if keyfile.read_to_string(&mut content).is_err() {
-                continue;
-            }
-
-            match KeyFile::decode(content) {
-                Ok(kf) => {
-                    let mut info = Vec::new();
-                    if kf.visible.is_none() || kf.visible.unwrap() || show_hidden {
-                        let is_hd = match kf.crypto {
-                            CryptoType::Core(_) => false,
-                            CryptoType::HdWallet(_) => true,
-                        };
-                        match kf.name {
-                            Some(name) => info.push(name),
-                            None => info.push("".to_string()),
-                        }
-
-                        match kf.description {
-                            Some(desc) => info.push(desc),
-                            None => info.push("".to_string()),
-                        }
-                        accounts.push((
-                            info[0].clone(),
-                            kf.address.to_string(),
-                            info[1].clone(),
-                            is_hd,
-                        ));
-                    }
-                }
-                Err(_) => info!("Invalid keystore file format for: {:?}", entry.file_name()),
-            }
-        }
-    }
-
-    Ok(accounts)
-}
-
-/// Hides account for given address from being listed
-///
-/// #Arguments
-/// addr - target address
-/// path - folder with keystore files
-///
-pub fn hide<P: AsRef<Path>>(addr: &Address, path: P) -> Result<bool, Error> {
-    let (p, mut kf) = KeyFile::search_by_address(addr, &path)?;
-
-    kf.visible = Some(false);
-    write(&kf, &p)?;
-
-    Ok(true)
-}
-
-/// Unhides account for given address from being listed
-///
-/// #Arguments
-/// addr - target address
-/// path - folder with keystore files
-///
-pub fn unhide<P: AsRef<Path>>(addr: &Address, path: P) -> Result<bool, Error> {
-    let (p, mut kf) = KeyFile::search_by_address(addr, &path)?;
-
-    kf.visible = Some(true);
-    write(&kf, &p)?;
-
-    Ok(true)
-}
-
-
-/// Creates filename for keystore file in format:
-/// `UTC--yyy-mm-ddThh-mm-ssZ--uuid`
-///
-/// # Arguments
-///
-/// * `uuid` - UUID for keyfile
-///
-fn generate_filename(uuid: &str) -> String {
-    format!("UTC--{}Z--{}", &util::timestamp(), &uuid)
 }
 
 #[cfg(test)]
